Fixed: -fixed wrong menus showing for trainer/trainee (For real this time)
Unfortunately, if we use a database query in a function, the function
may return before the query.  Even if we use oncomplete listener.  This
was leading to the menu error and we had to lose the isTrainee function.

Now the code works, but the query is only called once.  The init functions
for the menus take booleans instead of currentUser now.  Once the query
completes, then it calls the inits from its oncomplete listener, ensuring
that the real user type is found.  I found this out the hard way....lol